# app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Alarm:
    properties:
      acknowledged_at:
        type: string
      acknowledged_by:
        type: string
      agent_id:
        type: string
      config:
        $ref: '#/definitions/models.AlarmEscalationConfig'
      created_at:
        type: string
      current_level:
        $ref: '#/definitions/models.AlarmLevel'
      escalation_count:
        type: integer
      id:
        type: string
      is_acknowledged:
        type: boolean
      last_escalation:
        type: string
      message:
        type: string
      metadata:
        $ref: '#/definitions/models.JSONMap'
      priority:
        $ref: '#/definitions/models.NotificationPriority'
      project_id:
        type: string
      start_time:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.AlarmEscalationConfig:
    properties:
      audio_enabled:
        type: boolean
      broadcast_to_all:
        type: boolean
      escalation_interval:
        $ref: '#/definitions/time.Duration'
      initial_level:
        $ref: '#/definitions/models.AlarmLevel'
      max_level:
        $ref: '#/definitions/models.AlarmLevel'
      persist_until_acknowledged:
        type: boolean
      visual_enabled:
        type: boolean
    type: object
  models.AlarmLevel:
    enum:
    - soft
    - medium
    - loud
    - urgent
    - critical
    type: string
    x-enum-varnames:
    - AlarmLevelSoft
    - AlarmLevelMedium
    - AlarmLevelLoud
    - AlarmLevelUrgent
    - AlarmLevelCritical
  models.AlarmStats:
    properties:
      active_count:
        type: integer
      critical_count:
        type: integer
      total_today:
        type: integer
      unacknowledged_count:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  models.JSONMap:
    additionalProperties: true
    type: object
...
  models.NotificationPriority:
    enum:
    - low
    - normal
    - high
    - urgent
    - critical
    type: string
    x-enum-varnames:
    - NotificationPriorityLow
    - NotificationPriorityNormal
    - NotificationPriorityHigh
    - NotificationPriorityUrgent
    - NotificationPriorityCritical
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ticket Management System API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TMS API
  version: "1.0"
paths:
...
  /api/alarms/{alarm_id}/acknowledge:
    post:
      consumes:
      - application/json
      description: Acknowledge a specific alarm to mark it as seen/handled
      parameters:
      - description: Tenant ID
        in: header
        name: tenant_id
        required: true
        type: string
      - description: Project ID
        in: header
        name: project_id
        required: true
        type: string
      - description: Alarm ID
        in: path
        name: alarm_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Acknowledge alarm
      tags:
      - alarms
  /api/alarms/active:
    get:
      consumes:
      - application/json
      description: Retrieve all active alarms for a specific project
      parameters:
      - description: Tenant ID
        in: header
        name: tenant_id
        required: true
        type: string
      - description: Project ID
        in: header
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              alarms:
                items:
                  $ref: '#/definitions/models.Alarm'
                type: array
              total:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get active alarms
      tags:
      - alarms
  /api/alarms/notification-preferences:
    get:
      consumes:
      - application/json
      description: Retrieve notification preferences for alarm notifications
      parameters:
      - description: Tenant ID
        in: header
        name: tenant_id
        required: true
        type: string
      - description: Agent ID
        in: header
        name: agent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              agent_id:
                type: string
              browser_enabled:
                type: boolean
              desktop_enabled:
                type: boolean
              email_enabled:
                type: boolean
              escalation_timeout_minutes:
                type: integer
              sound_enabled:
                type: boolean
              tenant_id:
                type: string
              vibration_enabled:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notification preferences
      tags:
      - alarms
    put:
      consumes:
      - application/json
      description: Update notification preferences for alarm notifications
      parameters:
      - description: Tenant ID
        in: header
        name: tenant_id
        required: true
        type: string
      - description: Agent ID
        in: header
        name: agent_id
        required: true
        type: string
      - description: Notification preferences to update
        in: body
        name: preferences
        required: true
        schema:
          properties:
            browser_enabled:
              type: boolean
            desktop_enabled:
              type: boolean
            email_enabled:
              type: boolean
            escalation_timeout_minutes:
              type: integer
            sound_enabled:
              type: boolean
            vibration_enabled:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              agent_id:
                type: string
              message:
                type: string
              preferences:
                type: object
              tenant_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update notification preferences
      tags:
      - alarms
  /api/alarms/stats:
    get:
      consumes:
      - application/json
      description: Retrieve alarm statistics and metrics for a project
      parameters:
      - description: Tenant ID
        in: header
        name: tenant_id
        required: true
        type: string
      - description: Project ID
        in: header
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AlarmStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get alarm statistics
      tags:
      - alarms
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"